copy dlls from project root

+++ build-jom.bat
@@ -68,14 +68,18 @@ goto :return
 
 :do_test
 REM echo do testing
-pushd %builddir% >NUL
-dir /b /s *.dll > %dllsfile% 2>NUL
-for /f %%a in (%dllsfile%) do (
-    xcopy /Y %%a .\
+if not exist %curdir%\CMakeLists.txt goto :return
+if not exist %builddir%\CMakeCache.txt goto :return
+if exist %builddir%\CMakeCache.txt (
+    pushd %builddir% >NUL
+    dir /b /s ..\*.dll > %dllsfile% 2>NUL
+    for /f %%a in (%dllsfile%) do (
+        xcopy /Y %%a .\tests\ >NUL 2>NUL
+    )
+    ctest --output-on-failure
+    del /s /q %dllsfile% >NUL
+    popd >NUL
 )
-ctest --output-on-failure
-del /s /q %dllsfile% >NUL
-popd >NUL
 goto :return
 
 :do_clean
